Put say command in its own class
